tests/wpt/tests/css/css-display/reading-order/reading-order-items-invalid.html

<!DOCTYPE html>
<title>CSS Display: parsing reading-order-items with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-display-4/#reading-order-items">
<link rel="author" title="Di Zhang" href="mailto:dizhangg@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<body>
<script>
test_invalid_value('reading-order-items', 'auto');
test_invalid_value('reading-order-items', 'none');
test_invalid_value('reading-order-items', 'legacy');
test_invalid_value('reading-order-items', 'normal auto');
test_invalid_value('reading-order-items', 'flex-visual flex-flow');
test_invalid_value('reading-order-items', '100%');
test_invalid_value('reading-order-items', '10px');
</script>
</body>
